Fix truncateTables

+++ test/utils/truncate-tables.ts
@@ -17,6 +17,9 @@ export async function truncateTables() {
   const tables: { tablename: string }[] = await dataSource.query(
     `SELECT tablename FROM pg_tables WHERE schemaname = 'public'`,
   )
+  if (tables.length === 0) {
+    return
+  }
   const names = tables.map((row) => `"${row.tablename}"`).join(', ')
   const sql = `TRUNCATE TABLE ${names} CASCADE;`
   await dataSource.query(sql)
